% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacenter.R
\name{new_datacenter}
\alias{new_datacenter}
\title{Datacenter object}
\usage{
new_datacenter(vars)
}
\description{
An R6 class to represent source code/data associated with a political candidate
}
\section{Functions}{

Some text here

\describe{
\item{one}{description}
}
}

